+++ activitypub/adapters/lemmy/factories.py
@@ -1,10 +1,13 @@
 import factory
+from django.utils import timezone
 
 from activitypub.core.factories import (
     ActorFactory,
+    ObjectFactory,
     ReferenceFactory,
     UserFactory,
 )
+from activitypub.core.models import ObjectContext
 
 from . import models
 
@@ -38,6 +41,15 @@ class PostFactory(factory.django.DjangoModelFactory):
     reference = factory.SubFactory(ReferenceFactory)
     community = factory.SubFactory(CommunityFactory)
 
+    @factory.post_generation
+    def object_context(obj, create, extracted, **kw):
+        if create and not ObjectContext.objects.filter(reference=obj.reference).exists():
+            ObjectFactory(
+                reference=obj.reference,
+                type=ObjectContext.Types.PAGE,
+                published=timezone.now(),
+            )
+
     class Meta:
         model = models.Post
 
@@ -45,6 +57,15 @@ class PostFactory(factory.django.DjangoModelFactory):
 class CommentFactory(factory.django.DjangoModelFactory):
     reference = factory.SubFactory(ReferenceFactory)
     post = factory.SubFactory(PostFactory)
+
+    @factory.post_generation
+    def object_context(obj, create, extracted, **kw):
+        if create and not ObjectContext.objects.filter(reference=obj.reference).exists():
+            ObjectFactory(
+                reference=obj.reference,
+                type=ObjectContext.Types.NOTE,
+                published=timezone.now(),
+            )
 
     class Meta:
         model = models.Comment