+++ activitypub/adapters/lemmy/handlers.py
@@ -5,24 +5,27 @@ from django.db.models import F
 from django.db.models.signals import post_save
 from django.dispatch import receiver
 
-from activitypub.core.models import ActorContext, CollectionContext
+from activitypub.core.models import ActivityContext, ActorContext, CollectionContext
 from activitypub.core.signals import reference_loaded
 
-from .models import (
+from .models.aggregates import (
+    FollowerCount,
+    RankingScore,
+    ReactionCount,
+    SubmissionCount,
+    UserActivity,
+)
+from .models.core import (
     Comment,
-    CommentAggregates,
     Community,
-    CommunityAggregates,
     Person,
-    PersonAggregates,
     Post,
-    PostAggregates,
     Report,
     Site,
-    SiteReportedStatistics,
     UserProfile,
     UserSettings,
 )
+from .tasks import process_activity
 
 logger = logging.getLogger(__name__)
 User = get_user_model()
@@ -56,7 +59,13 @@ def on_person_created_create_aggregates_record(sender, **kw):
     person = kw["instance"]
 
     if kw["created"]:
-        PersonAggregates.objects.get_or_create(person=person)
+        FollowerCount.objects.get_or_create(reference=person.reference)
+        SubmissionCount.objects.get_or_create(
+            reference=person.reference, type=SubmissionCount.Types.POST
+        )
+        SubmissionCount.objects.get_or_create(
+            reference=person.reference, type=SubmissionCount.Types.COMMENT
+        )
 
 
 @receiver(post_save, sender=Person)
@@ -87,7 +96,8 @@ def on_community_created_create_aggregates_record(sender, **kw):
     community = kw["instance"]
 
     if kw["created"]:
-        CommunityAggregates.objects.get_or_create(community=community)
+        FollowerCount.objects.get_or_create(reference=community.reference)
+        UserActivity.objects.get_or_create(reference=community.reference)
 
 
 @receiver(post_save, sender=Post)
@@ -95,7 +105,12 @@ def on_post_created_create_aggregates_record(sender, **kw):
     post = kw["instance"]
 
     if kw["created"]:
-        PostAggregates.objects.create(post=post)
+        ReactionCount.objects.get_or_create(reference=post.reference)
+        SubmissionCount.objects.get_or_create(
+            reference=post.reference, type=SubmissionCount.Types.POST
+        )
+        for ranking_type in RankingScore.Types:
+            RankingScore.objects.get_or_create(type=ranking_type, reference=post.reference)
 
 
 @receiver(post_save, sender=Comment)
@@ -103,11 +118,18 @@ def on_comment_created_update_aggregates(sender, **kw):
     comment = kw["instance"]
 
     if kw["created"]:
-        CommentAggregates.objects.create(comment=comment)
-        PostAggregates.objects.update_or_create(
-            post=comment.post,
-            defaults={"comments": F("comments") + 1},
-            create_defaults={"comments": 1},
+        ReactionCount.objects.get_or_create(reference=comment.reference)
+        SubmissionCount.objects.get_or_create(
+            reference=comment.reference, type=SubmissionCount.Types.COMMENT
+        )
+        for ranking_type in RankingScore.Types:
+            RankingScore.objects.get_or_create(type=ranking_type, reference=comment.reference)
+
+        SubmissionCount.objects.update_or_create(
+            reference=comment.post.reference,
+            type=SubmissionCount.Types.COMMENT,
+            defaults={"replies": F("replies") + 1, "latest_reply": comment.as2.published},
+            create_defaults={"replies": 1, "type": SubmissionCount.Types.COMMENT},
         )
 
 
@@ -116,7 +138,10 @@ def on_site_created_create_aggregates_record(sender, **kw):
     site = kw["instance"]
 
     if kw["created"]:
-        SiteReportedStatistics.objects.create(site=site)
+        UserActivity.objects.get_or_create(reference=site.reference)
+        SubmissionCount.objects.get_or_create(
+            reference=site.reference, type=SubmissionCount.Types.POST
+        )
 
 
 @receiver(reference_loaded)