+++ activitypub/adapters/lemmy/tasks.py
@@ -4,7 +4,7 @@ from celery import shared_task
 
 from activitypub.core.contexts import AS2
 from activitypub.core.models import ActivityContext, ActorContext, Reference
-from activitypub.publishers import publish
+from activitypub.core.publishers import publish
 
 from .models.core import LemmyObject
 from .projections import lemmy_projection_selector